requirements.txt
fastapi[all]
aioredis==2.0.1
hiredis==2.0.0
pybit==5.6.2
pandas==2.2.1
sqlalchemy==2.0.28
pandas_ta==0.3.14b0
mplfinance==0.12.10b0
python-telegram-bot==21.0.1